🔨 refactor: tanstack query 의존성을 shared에서 web으로 이동
packages/shared에서 tanstack query를 제거하고 apps/web으로 이동하여
web 애플리케이션에서만 사용하도록 변경함.

- apps/web/package.json에 @tanstack/react-query, @tanstack/react-query-devtools 추가
- QueryProvider를 packages/shared에서 apps/web/components/providers로 이동
- app-provider.tsx의 QueryProvider import 경로를 로컬 경로로 변경
- packages/shared의 불필요한 의존성 및 설정 파일 제거
