+++ apps/web/package.json
@@ -11,7 +11,8 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.2",
-    "@triad/shared": "workspace:*",
+    "@tanstack/react-query": "^5.62.11",
+    "@tanstack/react-query-devtools": "^5.62.11",
     "@triad/ui": "workspace:*",
     "bcryptjs": "^3.0.2",
     "ky": "^1.2.3",
