+++ packages/shared/package.json
@@ -2,25 +2,12 @@
   "name": "@triad/shared",
   "version": "0.0.1",
   "private": true,
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch",
     "setup-dev": "tsc"
   },
-  "dependencies": {
-    "@tanstack/react-query": "^5.62.11",
-    "@tanstack/react-query-devtools": "^5.62.11"
-  },
   "devDependencies": {
-    "@triad/typescript-config": "workspace:*",
-    "@types/react": "^19.0.12",
-    "@types/react-dom": "^19.0.4",
-    "react": "^19.0.0",
-    "typescript": "^5.7.3"
-  },
-  "peerDependencies": {
-    "react": "^19.0.0"
+    "@triad/typescript-config": "workspace:*"
   }
 }
