--- packages/shared/src/index.ts
@@ -1,1 +1,0 @@
-export { QueryProvider } from "./providers/query-provider";
