--- packages/shared/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@triad/typescript-config/module",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "jsx": "react-jsx",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
